既存のテストを修正
