+++ test/integration/api/notifications_test.rb
@@ -13,12 +13,12 @@ class API::NotificationsTest < ActionDispatch::IntegrationTest
     assert_response :ok
   end
 
-  test 'GET /api/notifications/unread.json' do
-    get api_unread_index_path(format: :json)
+  test 'GET /api/notifications.json?status=unread' do
+    get api_notifications_path(status: 'unread', format: :json)
     assert_response :unauthorized
 
     token = create_token('hatsuno', 'testtest')
-    get api_unread_index_path(format: :json),
+    get api_notifications_path(status: 'unread', format: :json),
         headers: { 'Authorization' => "Bearer #{token}" }
     assert_response :ok
   end
